+++ quickdb/__init__.py
@@ -1,3 +1,3 @@
 from quickdb.main import QuickDB
 
-__all__ = ['QuickDB']
+__all__ = ["QuickDB"]