+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file. The format 
 ## Version 0.0.2 
 
 - Added support for `schema` and `table` in `Database` class
+- Support for logger added
+- Support for multiple tables in a database added
+- Support for drop table added
+- Support for drop database added
+- Support for `select` , `insert`, `update` and `delete` queries added
 
 ## [Initial Release]